pyproject.toml
[project]
name = "tts-impl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alias-free-torch>=0.0.6",
    "codespell>=2.4.1",
    "hydra-core>=1.3.2",
    "lightning==2.5.5",
    "mypy>=1.18.2",
    "numba>=0.62.1",
    "omegaconf>=2.3.0",
    "pyaudio>=0.2.14",
    "pyopenjtalk-plus>=0.4.1.post4",
    "pytest>=8.4.2",
    "pyworld>=0.3.5",
    "rich>=14.2.0",
    "rich-argparse>=1.7.1",
    "rotary-embedding-torch>=0.8.9",
    "ruff>=0.8.0",
    "safetensors>=0.6.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "soundfile>=0.13.1",
    "tensorboard>=2.20.0",
    "toml>=0.10.2",
    "torch-yin>=0.1.3",
    "torchfcpe>=0.0.4",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src/"]
testpaths = ["test"]

[tool.hatch.build.targets.wheel]
packages = ["src/tts_impl/"]

[tool.ruff]
line-length = 88
target-version = "py313"
src = ["src", "test", "scripts"]
exclude = ["third_party"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module level import not at top of file
    "E721",   # type comparison (common pattern in ML code)
    "E741",   # ambiguous variable name (l, I are common in math)
    "F401",   # imported but unused (for re-exports in __init__.py)
    "F841",   # local variable assigned but never used (common in tests)
    "B006",   # mutable default argument (common pattern with dataclasses)
    "B008",   # function call in default argument
    "B023",   # loop variable not bound (common in closures)
    "B028",   # stacklevel in warnings
    "B905",   # zip without strict
    "B026",   # star-arg unpacking after keyword argument
]

[tool.ruff.lint.isort]
known-first-party = ["tts_impl"]

[tool.mypy]
python_version = "3.13"
strict = true
mypy_path = ["stubs"]
files = ["src/", "test/", "scripts/"]
exclude = ["third_party/"]

[[tool.mypy.overrides]]
module = ["numpy.*", "torch.*", "torchaudio.*", "lightning.*", "hydra.*", "omegaconf.*", "scipy.*", "numba.*", "pyworld.*", "soundfile.*", "tqdm.*", "transformers.*", "safetensors.*", "tensorboard.*", "sklearn.*", "pyopenjtalk.*", "torchfcpe.*", "alias_free_activation.*", "rotary_embedding_torch.*", "toml.*", "torch_yin.*", "pyaudio.*", "alias_free_torch.*", "rich.*"]
ignore_missing_imports = true
